+++ Tarea 1/src/divide_and_concer/divide_and_conquer.cpp
@@ -1,16 +1,18 @@
 #include "divide_and_conquer.hpp"
+#include <algorithm>
 
 // Ordenar puntos por coordenada x
-bool compX(const std::pair<int, int> &p1, const std::pair<int, int> &p2) {
+bool compX(const std::pair<double, double> &p1, const std::pair<double, double> &p2) {
     return p1.first < p2.first;
 }
 
 // Ordenar puntos por coordenada y
-bool compY(const std::pair<int, int> &p1, const std::pair<int, int> &p2) {
+bool compY(const std::pair<double, double> &p1, const std::pair<double, double> &p2) {
     return p1.second < p2.second;
 }
 
-double minDistCentro(std::vector<std::pair<int, int>> &centro, double d) {
+double minDistCentro(std::vector<std::pair<double, double>> &centro, double d) {
+    // Ensure centro is a std::vector for compatibility with std::sort
     // Ordenar los puntos por coordenada y
     std::sort(centro.begin(), centro.end(), compY);
 
@@ -26,7 +28,7 @@ double minDistCentro(std::vector<std::pair<int, int>> &centro, double d) {
     return minDist;
 }
 
-double minDist(std::vector<std::pair<int, int>> &dots, int left, int right) {
+double minDist(std::vector<std::pair<double, double>> &dots, int left, int right) {
     // Caso base: si hay 2 o menos puntos, usamos fuerza bruta
     if (right - left <= 3) {
         double minDist = std::numeric_limits<double>::max();
@@ -50,7 +52,7 @@ double minDist(std::vector<std::pair<int, int>> &dots, int left, int right) {
     double d = std::min(dIzq, dDer);
 
     // Crear un vector para almacenar los puntos dentro de la distancia d
-    std::vector<std::pair<int, int>> centro;
+    std::vector<std::pair<double, double>> centro;
     for (int i = left; i < right; i++) {
         if (abs(dots[i].first - dots[mid].first) < d) {
             centro.push_back(dots[i]);
@@ -60,7 +62,7 @@ double minDist(std::vector<std::pair<int, int>> &dots, int left, int right) {
     return std::min(d, minDistCentro(centro, d));
 }
 
-double minDist_DQ(std::vector<std::pair<int, int>> &dots) {
+double minDist_DQ(std::vector<std::pair<double, double>> &dots) {
     // Ordenar los puntos por coordenada x
     std::sort(dots.begin(), dots.end(), compX);
